Encourage emoji to be drawn as images
Appends "Unicode Variant Selector 1" to the emoji property of weather, so that systems will default to drawing them as images instead of text, when able. Currently the heavy snow emoji is drawn as text in the discord end-of-game report on windows 10, similar to the Moist Talkers' ORB on blaseball.com.
This change is untested and superficial. But it's been every so slightly annoying me and I wanted to fix it.